Python_Module/requirements.txt
cobs
Pillow
RPi.GPIO
numpy
playsound
spidev
